build: add Release-Please configuration and hooks
Added [Release-Please](https://github.com/googleapis/release-please)
configuration to simplify releases to just clicking the merge button on
a PR. It bases the version off of your commits using
[Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
This also allows it to generate a changelog as well.

To ensure that commits following the convention, a git hook, using
[husky](https://typicode.github.io/husky/), has also
been added that prevents commits from being made unless they follow the
convention. Git hooks can be skipped using CLI flags to ensure full
conformity, a workflow has been added to ensure their correctness and to
run a quick check of the code.

Signed-off-by: Lucy Ekaterina <lucy@dragnof.pro>